✅ task: update rest-api todos with userId

+++ 03-admin-todo/src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from "@/actions/auth-actions";
 import prisma from "@/libs/prisma";
 import { NextResponse } from "next/server";
 import * as yup from "yup";
@@ -28,9 +29,16 @@ const postTodoSchema = yup.object({
 });
 
 export async function POST(request: Request) {
+  const user = await getUserSessionServer();
+  if (!user) {
+    return NextResponse.json({ messagge: "User unauthorized" }, { status: 401 });
+  }
+
   try {
     const { description, complete } = await postTodoSchema.validate(await request.json());
-    const createTodo = await prisma.todo.create({ data: { description, complete } });
+    const createTodo = await prisma.todo.create({
+      data: { description, complete, userId: user.id },
+    });
     return NextResponse.json({ data: createTodo });
   } catch (error) {
     return NextResponse.json(error, { status: 400 });
@@ -38,8 +46,12 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getUserSessionServer();
+  if (!user) {
+    return NextResponse.json({ messagge: "User unauthorized" }, { status: 401 });
+  }
   try {
-    await prisma.todo.deleteMany({ where: { complete: true } });
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });
     return NextResponse.json({ data: "Todos Deleted" });
   } catch (error) {
     return NextResponse.json(error, { status: 400 });
